Update README.md
CALCULATOR
An application used for basic arithmetis which is addition, subtraction, multiplication and division. 
It provides an area where the user inputs the first and the second number to be used then the operator. 
Example 
Enter number1: 40
Enter number2: 50
Choose operation( + or - or *  or /) :*
2000
Error Handling 
If a wrong input is made the application provides a retry for the user after displaying the message "Wrong input. Try Again".